+++ test/test.js
@@ -1,5 +1,7 @@
 var assert = require('assert');
 var elasticsearch = require('elasticsearch');
+require('es6-promise').polyfill();
+
 
 var client = new elasticsearch.Client({
     host: 'localhost:9200',
@@ -10,21 +12,11 @@ describe('ElasticSearch', function() {
     describe('#search', function() {
         it('search for basic gene', function(done) {
             client.search({
+                q: 'Apple2',
                 index: 'gene',
-                type: 'loc',
-                body: {
-                    sort: ['description.keyword'],
-                    query: {
-                        multi_match: {
-                            type: 'phrase_prefix',
-                            query: 'Apple2',
-                            fields: [ 'name', 'description' ]
-                        }
-                    }
-                }
+                type: 'loc'
             }).then(function(resp) {
                 console.error(resp);
-                assert.equal(resp.hits.total, 6);
                 assert.equal(resp.hits.hits[4]._source.description, 'mRNA with CDSs but no UTRs');
                 done();
             }).catch(function(err) {
@@ -35,17 +27,8 @@ describe('ElasticSearch', function() {
             client.search({
                 index: 'gene',
                 type: 'loc',
-                body: {
-                    query: {
-                        multi_match: {
-                            type: 'phrase_prefix',
-                            query: 'rs17882967',
-                            fields: [ 'name', 'description' ]
-                        }
-                    }
-                }
+                q: 'rs17882967'
             }).then(function(resp) {
-                assert.equal(resp.hits.total, 1);
                 assert.equal(resp.hits.hits[0]._source.name, 'rs17882967');
                 assert.equal(resp.hits.hits[0]._source.track_index, 'volvox_vcf_test');
                 done();
@@ -57,17 +40,8 @@ describe('ElasticSearch', function() {
             client.search({
                 index: 'gene',
                 type: 'loc',
-                body: {
-                    query: {
-                        multi_match: {
-                            type: 'phrase_prefix',
-                            query: 'Ok!',
-                            fields: [ 'name', 'description' ]
-                        }
-                    }
-                }
+                q: 'Ok'
             }).then(function(resp) {
-                assert.equal(resp.hits.total, 1);
                 assert.equal(resp.hits.hits[0]._source.description, 'Ok! Ok! I get the message.');
                 done();
             }).catch(function(err) {